安全告警

+++ interfaces/kits/native/want/src/want_params_wrapper.cpp
@@ -114,9 +114,9 @@ sptr<IWantParams> WantParamWrapper::Parse(const std::string &str)
     std::string key = "";
     int typeId = 0;
     if (ValidateStr(str)) {
-        for (unsigned int strnum = 0; strnum < str.size(); strnum++) {
+        for (size_t strnum = 0; strnum < str.size(); strnum++) {
             if (str[strnum] == '{' && key != "" && typeId == WantParams::VALUE_TYPE_WANTPARAMS) {
-                unsigned int num;
+                size_t num;
                 int count = 0;
                 for (num = strnum; num < str.size(); num++) {
                     if (str[num] == '{') {
@@ -167,9 +167,9 @@ WantParams WantParamWrapper::ParseWantParams(const std::string &str)
     if (!ValidateStr(str)) {
         return wantPaqrams;
     }
-    for (unsigned int strnum = 0; strnum < str.size(); strnum++) {
+    for (size_t strnum = 0; strnum < str.size(); strnum++) {
         if (str[strnum] == '{' && key != "" && typeId == WantParams::VALUE_TYPE_WANTPARAMS) {
-            unsigned int num;
+            size_t num;
             int count = 0;
             for (num = strnum; num < str.size(); num++) {
                 if (str[num] == '{') {
